fix(mfcc): adjust mfcc constant
A constant exists in the formula for converting between hz and the mel scale, which Meyda has had
set to 1125. The mel-scale is not standardised, but it seems that the most popular formula comes
from "Speech communication: human and machine" by Douglas O'Shaughnessy. That formula has the
constant as 1127. The issue mentioned below was opened, and suggests that we converge on
O'Shaughnessy's constant. This commit implements that change.

BREAKING CHANGE: A change in how Meyda performs frequency to mel scale conversion will result in
different mel-values between this and previous versions of Meyda.

fix #458